Hari slutändan byggt upp 2 sidor, en rental och en rent. just nu vet jag knappt vad jag gjort längre. Men rentals skulle vara som stöd för rent.

+++ Models/FilmCategory.cs
@@ -5,7 +5,7 @@ namespace Retro_grupp_g.Models;
 
 public partial class FilmCategory
 {
-    public ushort FilmId { get; set; }
+    public int FilmId { get; set; }
 
     public byte CategoryId { get; set; }
